Add package core, move some classes there
