+++ asistente_ladm_col/core/model_parser.py
