+++ asistente_ladm_col/gui/supplies/wiz_supplies_etl.py
@@ -21,28 +21,21 @@
 from qgis.PyQt.QtCore import (Qt,
                               QSettings,
                               QCoreApplication,
-                              pyqtSignal,
-                              QObject)
+                              pyqtSignal)
 from qgis.PyQt.QtWidgets import (QWizard,
                                  QSizePolicy,
                                  QGridLayout,
                                  QMessageBox)
-from qgis._core import QgsProject
-from qgis.core import (Qgis,
-                       QgsWkbTypes,
-                       QgsVectorLayer)
+from qgis.core import (Qgis)
 from qgis.gui import QgsMessageBar
 
-import processing
-
 from asistente_ladm_col.config.enums import EnumDbActionType
-from asistente_ladm_col.config.general_config import (LAYER,
-                                                      COLLECTED_DB_SOURCE,
+from asistente_ladm_col.config.general_config import (COLLECTED_DB_SOURCE,
                                                       SETTINGS_CONNECTION_TAB_INDEX,
-                                                      SUPPLIES_DB_SOURCE, 
-                                                      BLO_LIS_FILE_PATH,
-                                                      PREDIO_SANCION_FILE_PATH)
+                                                      SUPPLIES_DB_SOURCE)
 from asistente_ladm_col.config.help_strings import HelpStrings
+from asistente_ladm_col.core.supplies.etl_cobol import ETLCobol
+from asistente_ladm_col.core.supplies.etl_snc import ETLSNC
 from asistente_ladm_col.gui.dialogs.dlg_settings import SettingsDialog
 from asistente_ladm_col.gui.supplies.cobol_data_sources_widget import CobolDataSourceWidget
 from asistente_ladm_col.gui.supplies.snc_data_sources_widget import SNCDataSourceWidget
@@ -354,257 +347,3 @@ class SuppliesETLWizard(QWizard, WIZARD_UI):
         return True, ''
 
 
-class ETLSupplies(QObject):
-    CLASS_NAME = ""
-
-    def __init__(self, names, data_source_widget):
-        QObject.__init__(self)
-        self.names = names
-        self.data_source_widget = data_source_widget
-        self.logger = Logger()
-
-        self.gdb_path = self.data_source_widget.txt_file_path_gdb.text()
-
-        self.alphanumeric_file_paths = {}
-        self.gdb_layer_paths = dict()
-        self.layers = dict()
-
-        self.initialize_layers()
-
-    def initialize_layers(self):
-        raise NotImplementedError
-
-    def load_alphanumeric_layers(self):
-        raise NotImplementedError
-
-    def load_spatial_layers(self):
-        required_layers = ['R_TERRENO', 'U_TERRENO', 'R_SECTOR', 'U_SECTOR', 'R_VEREDA', 'U_MANZANA', 'U_BARRIO',
-                           'R_CONSTRUCCION', 'U_CONSTRUCCION', 'U_UNIDAD', 'R_UNIDAD',
-                           'U_NOMENCLATURA_DOMICILIARIA', 'R_NOMENCLATURA_DOMICILIARIA', 'U_PERIMETRO']
-
-        layer = QgsVectorLayer(self.gdb_path, 'layer name', 'ogr')
-
-        if not layer.isValid():
-            return False, QCoreApplication.translate(self.CLASS_NAME, "There were troubles loading the GDB.")
-
-        sublayers = layer.dataProvider().subLayers()
-
-        root = QgsProject.instance().layerTreeRoot()
-        gdb_group = root.addGroup(QCoreApplication.translate(self.CLASS_NAME, "GDB Supplies"))
-
-        for data in sublayers:
-            sublayer = data.split('!!::!!')[1]
-            if sublayer in required_layers:
-                layer = QgsVectorLayer(self.gdb_path + '|layername=' + sublayer, sublayer, 'ogr')
-                self.gdb_layer_paths[sublayer] = layer
-                QgsProject.instance().addMapLayer(layer, False)
-                gdb_group.addLayer(layer)
-
-        if len(self.gdb_layer_paths) != len(required_layers):
-            return False, QCoreApplication.translate(self.CLASS_NAME, "The GDB does not have the required layers!")
-
-        return True, ''
-
-    def run_etl_model(self, custom_feedback):
-        raise NotImplementedError
-
-
-class ETLCobol(ETLSupplies):
-    CLASS_NAME = "ETLCobol"
-
-    def __init__(self, names, data_source_widget):
-        ETLSupplies.__init__(self, names, data_source_widget)
-
-    def initialize_layers(self):
-        self.layers = {
-            self.names.GC_PARCEL_T: {'name': self.names.GC_PARCEL_T, 'geometry': None, LAYER: None},
-            self.names.GC_OWNER_T: {'name': self.names.GC_OWNER_T, 'geometry': None, LAYER: None},
-            self.names.GC_ADDRESS_T: {'name': self.names.GC_ADDRESS_T, 'geometry': QgsWkbTypes.LineGeometry, LAYER: None},
-            self.names.GC_BUILDING_UNIT_T: {'name': self.names.GC_BUILDING_UNIT_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_BUILDING_T: {'name': self.names.GC_BUILDING_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_PLOT_T: {'name': self.names.GC_PLOT_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_RURAL_DIVISION_T: {'name': self.names.GC_RURAL_DIVISION_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_URBAN_SECTOR_T: {'name': self.names.GC_URBAN_SECTOR_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_RURAL_SECTOR_T: {'name': self.names.GC_RURAL_SECTOR_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_PERIMETER_T: {'name': self.names.GC_PERIMETER_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_BLOCK_T: {'name': self.names.GC_BLOCK_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_NEIGHBOURHOOD_T: {'name': self.names.GC_NEIGHBOURHOOD_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_COMMISSION_BUILDING_T: {'name': self.names.GC_COMMISSION_BUILDING_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_COMMISSION_PLOT_T: {'name': self.names.GC_COMMISSION_PLOT_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_COMMISSION_BUILDING_UNIT_T: {'name': self.names.GC_COMMISSION_BUILDING_UNIT_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None}
-        }
-
-    def load_alphanumeric_layers(self):
-        self.alphanumeric_file_paths = {
-            'blo': self.data_source_widget.txt_file_path_blo.text().strip(),
-            'uni': self.data_source_widget.txt_file_path_uni.text().strip(),
-            'ter': self.data_source_widget.txt_file_path_ter.text().strip(),
-            'pro': self.data_source_widget.txt_file_path_pro.text().strip()
-        }
-
-        root = QgsProject.instance().layerTreeRoot()
-        lis_group = root.addGroup(QCoreApplication.translate(self.CLASS_NAME, "LIS Supplies"))
-
-        for name in self.alphanumeric_file_paths:
-            uri = 'file:///{}?type=csv&delimiter=;&detectTypes=yes&geomType=none&subsetIndex=no&watchFile=no'.format(self.alphanumeric_file_paths[name])
-            layer = QgsVectorLayer(uri, name, 'delimitedtext')
-            if layer.isValid():
-                self.alphanumeric_file_paths[name] = layer
-                QgsProject.instance().addMapLayer(layer, False)
-                lis_group.addLayer(layer)
-            else:
-                if name == 'blo':
-                    # BLO is kind of optional, if it is not given, we pass a default one
-                    uri = 'file:///{}?type=csv&delimiter=;&detectTypes=yes&geomType=none&subsetIndex=no&watchFile=no'.format(BLO_LIS_FILE_PATH)
-                    layer = QgsVectorLayer(uri, name, 'delimitedtext')
-                    self.alphanumeric_file_paths[name] = layer
-                    QgsProject.instance().addMapLayer(layer, False)
-                    lis_group.addLayer(layer)
-                else:
-                    return False, QCoreApplication.translate(self.CLASS_NAME, "There were troubles loading the LIS file called '{}'.".format(name))
-
-        return True, ''
-
-    def run_etl_model(self, custom_feedback):
-        self.logger.info(__name__, "Running ETL-Cobol model...")
-        processing.run("model:ETL-model-supplies",
-                       {'barrio': self.gdb_layer_paths['U_BARRIO'],
-                        'gcbarrio': self.layers[self.names.GC_NEIGHBOURHOOD_T][LAYER],
-                        'gccomisionconstruccion': self.layers[self.names.GC_COMMISSION_BUILDING_T][LAYER],
-                        'gccomisionterreno': self.layers[self.names.GC_COMMISSION_PLOT_T][LAYER],
-                        'gcconstruccion': self.layers[self.names.GC_BUILDING_T][LAYER],
-                        'gcdireccion': self.layers[self.names.GC_ADDRESS_T][LAYER],
-                        'gcmanzana': self.layers[self.names.GC_BLOCK_T][LAYER],
-                        'gcperimetro': self.layers[self.names.GC_PERIMETER_T][LAYER],
-                        'gcpropietario': self.layers[self.names.GC_OWNER_T][LAYER],
-                        'gcsector': self.layers[self.names.GC_RURAL_SECTOR_T][LAYER],
-                        'gcsectorurbano': self.layers[self.names.GC_URBAN_SECTOR_T][LAYER],
-                        'gcterreno': self.layers[self.names.GC_PLOT_T][LAYER],
-                        'gcunidad': self.layers[self.names.GC_BUILDING_UNIT_T][LAYER],
-                        'gcunidadconstruccioncomision': self.layers[self.names.GC_COMMISSION_BUILDING_UNIT_T][LAYER],
-                        'gcvereda': self.layers[self.names.GC_RURAL_DIVISION_T][LAYER],
-                        'inputblo': self.alphanumeric_file_paths['blo'],
-                        'inputconstruccion': self.gdb_layer_paths['R_CONSTRUCCION'],
-                        'inputmanzana': self.gdb_layer_paths['U_MANZANA'],
-                        'inputperimetro': self.gdb_layer_paths['U_PERIMETRO'],
-                        'inputpro': self.alphanumeric_file_paths['pro'],
-                        'inputrunidad': self.gdb_layer_paths['R_UNIDAD'],
-                        'inputsector': self.gdb_layer_paths['R_SECTOR'],
-                        'inputter': self.alphanumeric_file_paths['ter'],
-                        'inputterreno': self.gdb_layer_paths['R_TERRENO'],
-                        'inputuconstruccion': self.gdb_layer_paths['U_CONSTRUCCION'],
-                        'inputuni': self.alphanumeric_file_paths['uni'],
-                        'inputusector': self.gdb_layer_paths['U_SECTOR'],
-                        'inpututerreno': self.gdb_layer_paths['U_TERRENO'],
-                        'inputuunidad': self.gdb_layer_paths['U_UNIDAD'],
-                        'inputvereda': self.gdb_layer_paths['R_VEREDA'],
-                        'ouputlayer': self.layers[self.names.GC_PARCEL_T][LAYER],
-                        'rnomenclatura': self.gdb_layer_paths['R_NOMENCLATURA_DOMICILIARIA'],
-                        'unomenclatura': self.gdb_layer_paths['U_NOMENCLATURA_DOMICILIARIA']},
-                       feedback=custom_feedback)
-        self.logger.info(__name__, "ETL-Cobol model finished.")
-
-
-class ETLSNC(ETLSupplies):
-    CLASS_NAME = "ETLSNC"
-
-    def __init__(self, names, data_source_widget):
-        ETLSupplies.__init__(self, names, data_source_widget)
-
-    def initialize_layers(self):
-        self.layers = {
-            self.names.GC_PARCEL_T: {'name': self.names.GC_PARCEL_T, 'geometry': None, LAYER: None},
-            self.names.GC_OWNER_T: {'name': self.names.GC_OWNER_T, 'geometry': None, LAYER: None},
-            self.names.GC_HP_CONDOMINIUM_DATA_T: {'name': self.names.GC_HP_CONDOMINIUM_DATA_T, 'geometry': None, LAYER: None},
-            self.names.GC_COPROPERTY_T: {'name': self.names.GC_COPROPERTY_T, 'geometry': None, LAYER: None},
-            self.names.GC_ADDRESS_T: {'name': self.names.GC_ADDRESS_T, 'geometry': QgsWkbTypes.LineGeometry, LAYER: None},
-            self.names.GC_BUILDING_UNIT_T: {'name': self.names.GC_BUILDING_UNIT_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_BUILDING_T: {'name': self.names.GC_BUILDING_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_PLOT_T: {'name': self.names.GC_PLOT_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_RURAL_DIVISION_T: {'name': self.names.GC_RURAL_DIVISION_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_URBAN_SECTOR_T: {'name': self.names.GC_URBAN_SECTOR_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_RURAL_SECTOR_T: {'name': self.names.GC_RURAL_SECTOR_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_PERIMETER_T: {'name': self.names.GC_PERIMETER_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_BLOCK_T: {'name': self.names.GC_BLOCK_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_NEIGHBOURHOOD_T: {'name': self.names.GC_NEIGHBOURHOOD_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_COMMISSION_BUILDING_T: {'name': self.names.GC_COMMISSION_BUILDING_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_COMMISSION_PLOT_T: {'name': self.names.GC_COMMISSION_PLOT_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None},
-            self.names.GC_COMMISSION_BUILDING_UNIT_T: {'name': self.names.GC_COMMISSION_BUILDING_UNIT_T, 'geometry': QgsWkbTypes.PolygonGeometry, LAYER: None}
-        }
-
-    def load_alphanumeric_layers(self):
-        self.alphanumeric_file_paths = {
-            'predio_sancion': self.data_source_widget.txt_file_path_predio_sancion.text().strip(),
-            'predio': self.data_source_widget.txt_file_path_predio.text().strip(),
-            'direccion': self.data_source_widget.txt_file_path_direccion.text().strip(),
-            'unidad_construccion': self.data_source_widget.txt_file_path_uni.text().strip(),
-            'persona': self.data_source_widget.txt_file_path_persona.text().strip(),
-            'ficha_matriz': self.data_source_widget.txt_file_path_ficha_m.text().strip(),
-            'ficha_matriz_predio': self.data_source_widget.txt_file_path_ficha_m_predio.text().strip(),
-        }
-
-        root = QgsProject.instance().layerTreeRoot()
-        alphanumeric_group = root.addGroup(QCoreApplication.translate(self.CLASS_NAME, "SNC Alphanumeric Supplies"))
-
-        for name in self.alphanumeric_file_paths:
-            uri = 'file:///{}?type=csv&delimiter=,&detectTypes=yes&geomType=none&subsetIndex=no&watchFile=no'.format(self.alphanumeric_file_paths[name])
-            layer = QgsVectorLayer(uri, name, 'delimitedtext')
-            if layer.isValid():
-                self.alphanumeric_file_paths[name] = layer
-                QgsProject.instance().addMapLayer(layer, False)
-                alphanumeric_group.addLayer(layer)
-            else:
-                if name == 'predio_sancion':
-                    # predio_sancion is kind of optional, if it is not given, we pass a default one
-                    uri = 'file:///{}?type=csv&delimiter=,&detectTypes=yes&geomType=none&subsetIndex=no&watchFile=no'.format(PREDIO_SANCION_FILE_PATH)
-                    layer = QgsVectorLayer(uri, name, 'delimitedtext')
-                    self.alphanumeric_file_paths[name] = layer
-                    QgsProject.instance().addMapLayer(layer, False)
-                    alphanumeric_group.addLayer(layer)
-                else:
-                    return False, QCoreApplication.translate(self.CLASS_NAME, "There were troubles loading the CSV file called '{}'.".format(name))
-
-        return True, ''
-
-    def run_etl_model(self, custom_feedback):
-        self.logger.info(__name__, "Running ETL-SNC model...")
-        processing.run("model:ETL_SNC", 			
-                       {'barrio': self.gdb_layer_paths['U_BARRIO'],
-                        'fichamatriz': self.alphanumeric_file_paths['ficha_matriz'],
-                        'fichamatrizpredio': self.alphanumeric_file_paths['ficha_matriz_predio'],
-                        'gcbarrio': self.layers[self.names.GC_NEIGHBOURHOOD_T][LAYER],
-                        'gccomisionesconstruccion': self.layers[self.names.GC_COMMISSION_BUILDING_T][LAYER],
-                        'gccomisionesterreno': self.layers[self.names.GC_COMMISSION_PLOT_T][LAYER],
-                        'gccomisionesunidadconstruccion': self.layers[self.names.GC_COMMISSION_BUILDING_UNIT_T][LAYER],
-                        'gcconstruccion': self.layers[self.names.GC_BUILDING_T][LAYER],
-                        'gccopropiedad': self.layers[self.names.GC_COPROPERTY_T][LAYER],
-                        'gcdireccion': self.layers[self.names.GC_ADDRESS_T][LAYER],
-                        'gcmanzana': self.layers[self.names.GC_BLOCK_T][LAYER],
-                        'gcperimetro': self.layers[self.names.GC_PERIMETER_T][LAYER],
-                        'gcpredio': self.layers[self.names.GC_PARCEL_T][LAYER],
-                        'gcpropiedadhorizontal': self.layers[self.names.GC_HP_CONDOMINIUM_DATA_T][LAYER],
-                        'gcpropietario': self.layers[self.names.GC_OWNER_T][LAYER],
-                        'gcsectorrural': self.layers[self.names.GC_RURAL_SECTOR_T][LAYER],
-                        'gcsectorurbano': self.layers[self.names.GC_URBAN_SECTOR_T][LAYER],
-                        'gcterreno': self.layers[self.names.GC_PLOT_T][LAYER],
-                        'gcunidadconstruccion': self.layers[self.names.GC_BUILDING_UNIT_T][LAYER],
-                        'gcvereda': self.layers[self.names.GC_RURAL_DIVISION_T][LAYER],
-                        'manzana': self.gdb_layer_paths['U_MANZANA'],
-                        'persona': self.alphanumeric_file_paths['persona'],
-                        'predio': self.alphanumeric_file_paths['predio'],
-                        'prediodireccion': self.alphanumeric_file_paths['direccion'],
-                        'rconstruccion': self.gdb_layer_paths['R_CONSTRUCCION'],
-                        'rnomemclatura': self.gdb_layer_paths['R_NOMENCLATURA_DOMICILIARIA'],	
-                        'rperimetro': self.gdb_layer_paths['U_PERIMETRO'],
-                        'rsector': self.gdb_layer_paths['R_SECTOR'],
-                        'rterreno': self.gdb_layer_paths['R_TERRENO'],
-                        'runidad': self.gdb_layer_paths['R_UNIDAD'],
-                        'uconstruccion': self.gdb_layer_paths['U_CONSTRUCCION'],
-                        'unidadconstruccion': self.alphanumeric_file_paths['unidad_construccion'],
-                        'unomenclatura': self.gdb_layer_paths['U_NOMENCLATURA_DOMICILIARIA'],
-                        'usector': self.gdb_layer_paths['U_SECTOR'],
-                        'uterreno': self.gdb_layer_paths['U_TERRENO'],
-                        'uunidad': self.gdb_layer_paths['U_UNIDAD'],
-                        'vereda': self.gdb_layer_paths['R_VEREDA']},
-                        feedback=custom_feedback)
-        self.logger.info(__name__, "ETL-SNC model finished.")
